refactor: convert pets store to TypeScript
- Rename pets.js -> pets.ts with full type annotations
- Add Tab type ('pets' | 'editor') for activeTab
- Type all method parameters and store generics
- Normalize error handling with instanceof Error checks
- Remove unnecessary async from selectPet (was sync operation)
